mysite/requirements-sub.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
boto3==1.24.4
botocore==1.27.4
certifi==2022.5.18.1
cffi==1.15.0
chardet==3.0.4
colorama==0.4.4
cryptography==37.0.2
Cython==0.29.30
Django==4.0.5
django-crontab==0.7.1
djangorestframework==3.13.1
filelock==3.7.1
flatbuffers==2.0
gluonnlp==0.10.0
graphviz==0.8.4
h11==0.13.0
huggingface-hub==0.7.0
idna==2.6
jmespath==1.0.0
kobert @ git+https://git@github.com/SKTBrain/KoBERT.git@e1f2f37055e7460d8427f6912579c0162cb69831
mxnet==1.7.0.post2
mysqlclient==2.1.0
numpy==1.22.4
onnxruntime==1.8.0
outcome==1.1.0
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.2
protobuf==4.21.1
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.18.4
s3transfer==0.6.0
sentencepiece==0.1.96
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
transformers==4.19.2
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
wsproto==1.1.0
